\name{getAppender}
\alias{getAppender}
\alias{addAppender}
\title{ Manage loggers within 'futile.logger' }
\description{
}
\usage{
getAppender(name)

addAppender(name, threshold, appender, layout)
}
\arguments{
  \item{name}{ }
  \item{threshold}{ }
  \item{appender}{ }
  \item{layout}{ }
}
\details{
}
\value{
}
\author{ Brian Lee Yung Rowe }
\examples{
}
\keyword{ data }
