\name{logger.options}
\alias{logger.options}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{ Constants for 'futile.logger' }
\description{
  The logging configuration is managed by 'logger.options', a function generated
  by OptionsManager within 'futile.options'. Other constants defined include
  the log levels.
}
\details{
  Constants provided by 'futile.logger'.
}
\author{ Brian Lee Yung Rowe }
\seealso{ \code{futile.options} }
\keyword{ data }
