\name{getLayout}
\alias{getLayout}
\alias{addLayout}
\title{ Manage loggers within 'futile.logger' }
\description{
}
\usage{
getLayout(name)

addLayout(name, threshold, appender, layout)
}
\arguments{
  \item{name}{ }
  \item{threshold}{ }
  \item{appender}{ }
  \item{layout}{ }
}
\details{
}
\value{
}
\author{ Brian Lee Yung Rowe }
\examples{
}
\keyword{ data }
