% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pylogger.R
\name{logger.trace}
\alias{logger.debug}
\alias{logger.error}
\alias{logger.fatal}
\alias{logger.info}
\alias{logger.trace}
\alias{logger.warn}
\title{Python-Style Logging STatements}
\usage{
logger.trace(..., name = "ROOT")
}
\arguments{
\item{name}{ignored -- included to match \code{flog.debug(...)} signature}

\item{\dots}{arguments to \code{flog.debug(...)} style log statement}
}
\value{
No return value.
}
\description{
After initializing the level-specific namespaces with \code{logger.setup(...)}
these statements can be used in exactly the same way as the equivalent \code{flog.debug(...)} style statements.

Log statements will end up in all appropriate log files, emulating Python-style logging.
}
\examples{
\dontrun{
logger.setup(debugLog='debug.log', infoLog='info.log', errorLog='error.log')
logger.trace('trace statement')
logger.debug('debug statement')
logger.info('info statement')
logger.warn('warn statement')
logger.error('error statement')
logger.fatal('fatal statement')
}
}
\author{
Jonathan Callahan
}
\seealso{
\code{\link{logger.setup}}
}
\keyword{data}

