% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{prepare_arg}
\alias{prepare_arg}
\title{Provide basic parsing for layout string}
\usage{
prepare_arg(x)
}
\arguments{
\item{x}{The argument to prepare}
}
\description{
Return name of argument if arg is empty. Otherwise return the value.
}
