\name{getLogger}
\alias{getLogger}
\alias{addLogger}
\title{ Manage loggers within 'futile.logger' }
\description{
}
\usage{
getLogger(name)

addLogger(name, threshold, appender, layout)
}
\arguments{
  \item{name}{ }
  \item{threshold}{ }
  \item{appender}{ }
  \item{layout}{ }
}
\details{
}
\value{
}
\author{ Brian Lee Yung Rowe }
\examples{
}
\keyword{ data }
