% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{logger.options}
\alias{logger.options}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{Constants for 'futile.logger'}
\usage{
logger.options(..., simplify = FALSE, update = list())
}
\arguments{
\item{...}{TODO}

\item{simplify}{TODO}

\item{update}{TODO}
}
\description{
Log level constants and the logger options.
}
\details{
The logging configuration is managed by 'logger.options', a function
generated by OptionsManager within 'futile.options'.
}
\seealso{
\code{futile.options}
}
\author{
Brian Lee Yung Rowe
}
\keyword{data}
