\name{logger.options}
\alias{DEBUG}
\alias{ERROR}
\alias{FATAL}
\alias{INFO}
\alias{logger.options}
\alias{TRACE}
\alias{WARN}
\title{Constants for 'futile.logger'}
\usage{
  logger.options(..., simplify = FALSE, update = list())
}
\arguments{
  \item{...}{TODO}

  \item{simplify}{TODO}

  \item{update}{TODO}
}
\description{
  Log level constants and the logger options.
}
\details{
  The logging configuration is managed by 'logger.options',
  a function generated by OptionsManager within
  'futile.options'.
}
\author{
  Brian Lee Yung Rowe
}
\seealso{
  \code{futile.options}
}
\keyword{data}

